(fix): correct time output in rest api

+++ output/raw/db_queries.php
@@ -76,7 +76,7 @@ class QM_Output_Raw_DB_Queries extends QM_Output_Raw {
 
 		return array(
 			'total'   => $db->total_qs,
-			'time'    => (float) number_format_i18n( $db->total_time, 4 ),
+			'time'    => number_format_i18n( $db->total_time, 4 ),
 			'queries' => $output,
 		);
 	}
@@ -86,7 +86,7 @@ class QM_Output_Raw_DB_Queries extends QM_Output_Raw {
 
 		$output['i']    = ++$this->query_row;
 		$output['sql']  = $row['sql'];
-		$output['time'] = (float) number_format_i18n( $row['ltime'], 4 );
+		$output['time'] = number_format_i18n( $row['ltime'], 4 );
 
 		if ( isset( $row['trace'] ) ) {
 			$stack = array();
